Architectural changes which remove the direct integration with Appenders. Instead, the MDC is more heavily used to retrieve the additional LogMachine metadata, including the full JSON format (which is lazily initialized).
In general, the approach to Groups has been imporoved and changed to
Topics and TopicBrokers. The brokers are separate from the normally
configured appenders, and it is up to the user to configure them from
within their application, including subscribing to topics.

The deduplicating filter is the one bit of ugliness, necessary if an
appender is subscribed to multiple topics.